+++ Program/data/objects/Award.cs
@@ -2,13 +2,13 @@
 
 namespace Rejestracja.Data.Objects {
     class Award {
-        public long id;
+        public int id;
         public String title;
-        public long displayOrder;
+        public int displayOrder;
 
         public const int TITLE_MAX_LENGTH = 256;
 
-        public Award(long id, String title, long displayOrder) {
+        public Award(int id, String title, int displayOrder) {
             this.id = id;
             this.title = title;
             this.displayOrder = displayOrder;
